[SAFE-474] move extract fields method
