+++ src/utils/form/index.ts
@@ -1,3 +1,5 @@
 export * from './removeField';
 export * from './addField';
 export * from './replaceField';
+export * from './findDuplicateFields';
+export * from './extractFields';
